+++ lib/rabbit/event_handler.rb
@@ -10,8 +10,8 @@ class Rabbit::EventHandler
 
     def inherited(subclass)
       super
-      subclass.ignore_queue_conversion = false
-      subclass.additional_job_configs = {}
+      subclass.ignore_queue_conversion ||= false
+      subclass.additional_job_configs ||= {}
       subclass.queue ||= queue
     end